Fixed Static Code Compilation
